添加文件修改时间判断

+++ tomcat_logbak.py
@@ -4,6 +4,7 @@ import os
 import datetime
 import shutil
 import sys
+import time
 
 if len(sys.argv) < 3:
     print("Usage:",sys.argv[0], "[DIR tomcat0 tomcat1...]")
@@ -25,10 +26,12 @@ for project in sys.argv[2:]:
         print ('Successfully created directory:', targetdir)
 
     for file in os.listdir(sourcedir):
-        if os.path.isfile(sourcedir + file) and file != "catalina.out":
+        #遍历源目录,找出文件,且文件的修改时间小于上个月最后一秒
+        if os.path.isfile(sourcedir + file) and file != "catalina.out" and os.path.getmtime(sourcedir + file) < time.mktime(datetime.date(day=1, month= datetime.datetime.today().month, year= datetime.datetime.today().year).timetuple()):
+            print(file)
             try:
                 shutil.move(sourcedir + file, targetdir)
             except shutil.Error:
-                print(sys.exc_info()[0])
+                print(sys.exc_info()[1])
             else:
                 print("Successfully moved file:",sourcedir + file)
